Mac OS X build fixes, Windows compile warning fixes, and miscellaneous cleanup.
git-svn-id: https://csp.svn.sourceforge.net/svnroot/csp/trunk@1776 7152ba15-5bd8-0310-b828-e896570a432b

+++ CSP/csp/cspsim/f16/F16Engine.cpp
@@ -293,7 +293,7 @@ void F16Engine::updatePrestart(double dt) {
 		return;
 	}
 	const double rpm = b_RPM->value();
-	const double rpm_factor = powf(rpm, 2.1);
+	const double rpm_factor = pow(rpm, 2.1);
 	const double drive = m_Drive * rpm_factor * getAltitudeFactor();
 	driveEngine(drive, dt);
 	// TODO jfs
